feat: add ConsumeStream for Docker build output handling
This improves support for output handling during docker build.
Specifically, this change makes sure that the docker image is
always built before running any of the other tests (TestMain).
If there is a build failure, this will always be reported,
while if run without the -v flag and successful docker build,
no output is emitted. If the -v flag is provided, docker build
output is also emitted, both as part of the TestMain run, and
as part of the TestBuild test.